requirements.txt
PyQt6>=6.6.0
sounddevice>=0.4.6
numpy>=1.26.0
imageio-ffmpeg>=0.5.0
python-mpv>=1.0.0
